Complete the unit test cases for service layer.

+++ order-management-data/src/main/java/org/neptrueworks/ordermanagement/data/reposition/IDataCountable.java
@@ -5,6 +5,6 @@ import org.neptrueworks.ordermanagement.data.maneuvering.IDataMappable;
 
 public interface IDataCountable <TEntity extends IDataEntitizable<TIdentifier>, TIdentifier extends Comparable<TIdentifier>>
         extends IDataMappable {
-    int countScalar(TIdentifier id);
-    int countAll();
+    Long countScalar(TIdentifier id);
+    Long countAll();
  }
